backtracking function done

+++ include/Gestor.h
@@ -35,6 +35,6 @@ public:
      */
      void MostrarNodes();
      void VerAdj();
-     void BackTracking(int n, int currentNode, int count, int currentCost, int &bestCost);
+     void BackTracking(int n, int currentNode, int count, int currentCost, double &bestCost);
 };
 #endif //DA_PROJETO_1_GESTOR_H
